pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentflow"
version = "0.1.0"
description = "Phase 1 YAML agent schema validation CLI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111,<1",
  "httpx>=0.27,<1",
  "Jinja2>=3.1,<4",
  "pydantic>=2.7,<3",
  "psycopg[binary]>=3.1,<4",
  "PyYAML>=6.0,<7",
  "SQLAlchemy>=2.0,<3",
  "uvicorn[standard]>=0.30,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
  "pytest-cov>=5,<7",
]

[project.scripts]
createAgent = "agentflow.cli:main"
agentflow = "agentflow.cli:main"

[tool.setuptools.packages.find]
include = ["agentflow*"]

[tool.setuptools.package-data]
agentflow = ["sql/migrations/*.sql", "viewer/templates/*.html", "viewer/static/*.css"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
markers = [
  "db: integration tests that require TEST_DATABASE_URL-backed PostgreSQL",
]
